% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarisePrediction.R
\name{add.map.legend}
\alias{add.map.legend}
\title{Add legend to map}
\usage{
add.map.legend(
  breaks,
  cols,
  main = "log(gc/L)",
  dis.bks = c(`6` = 6, `8` = 8, `10` = 10, `12` = 12)
)
}
\description{
Add legend to map
}
