% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateLSOA.R
\name{aggregate_lsoa}
\alias{aggregate_lsoa}
\title{aggregate LSOA concentration to the given geographical level for a given area at a given time point}
\usage{
aggregate_lsoa(area, m, agg.index, pw = NULL, method = "fast")
}
\description{
aggregate LSOA concentration to the given geographical level for a given area at a given time point
}
