% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarisePrediction.R
\name{bi.legend}
\alias{bi.legend}
\title{Construct biscale}
\usage{
bi.legend(
  cols,
  xlab,
  ylab,
  dis.bks = c(0, 0.8, 1),
  counts = NULL,
  size.lab,
  size.bks
)
}
\description{
Construct biscale
}
