% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{extract.1d.sptm.interactions}
\alias{extract.1d.sptm.interactions}
\title{Extract space-time interaction samples for a given site or a given time point}
\usage{
extract.1d.sptm.interactions(
  interactions,
  itm = NULL,
  isp = NULL,
  site.index,
  time.index
)
}
\description{
Extract space-time interaction samples for a given site or a given time point
}
