% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{extract.joint.hyperpar}
\alias{extract.joint.hyperpar}
\title{Extract hyperparameters from the posterior draws of the INLA approximate joint posterior}
\usage{
extract.joint.hyperpar(samples, para)
}
\description{
Extract hyperparameters from the posterior draws of the INLA approximate joint posterior
}
