% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{extract.posterior.samples.from.joint}
\alias{extract.posterior.samples.from.joint}
\title{Obtain posterior samples from the INLA approximate joint posterior}
\usage{
extract.posterior.samples.from.joint(
  nsims,
  fits,
  dm,
  A.indexs.spde,
  fit.object = TRUE,
  region_effect
)
}
\description{
Note: The smoothing parameter, nu, in the Matérn covariance function is fixed at 1. (this note is no longer relevant)
This comes from the setting alpha=2 in INLA::inla.spde2.pcmatern (called in spde.make.A in the file 03_run_INLA.R).
See the note there for explanation.
}
