% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlaFitting.R
\name{fit_inla}
\alias{fit_inla}
\title{fit inla model from data, covariates, and predictor.}
\usage{
fit_inla(ww.weekly, covars, y, priors, mesh = NULL, region_effect)
}
\description{
fit inla model from data, covariates, and predictor.
}
