% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarisePrediction.R
\name{forestplot}
\alias{forestplot}
\title{for producing a forest plot to visualise level correlation between WW and debiased prevalence}
\usage{
forestplot(
  d,
  xlab = "week start",
  main = "Correlation of level between log(gc/L) and debiased prevalence",
  ylab = "Correlation",
  week.ids
)
}
\description{
for producing a forest plot to visualise level correlation between WW and debiased prevalence
}
