% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlaFitting.R
\name{form.fitdata.stk}
\alias{form.fitdata.stk}
\title{y_column can be 'log_gc' (for the original data)
or 'log_gcX' where X=1,...,nsims, one of the simulated datasets
with LOD entries simulated from U(0,L)}
\usage{
form.fitdata.stk(
  fitdata,
  A.indexs,
  X = NULL,
  y_column = "log_e_gc",
  region_effect
)
}
\description{
y_column can be 'log_gc' (for the original data)
or 'log_gcX' where X=1,...,nsims, one of the simulated datasets
with LOD entries simulated from U(0,L)
}
