% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{get.geo.lookups}
\alias{get.geo.lookups}
\title{Construct lookup tables between LSOA11 and various other admin geographies (LAD21, CCG21, RGN21 and CTRY21)}
\usage{
get.geo.lookups()
}
\description{
Construct lookup tables between LSOA11 and various other admin geographies (LAD21, CCG21, RGN21 and CTRY21)
}
