% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{get.LSOA.covariates}
\alias{get.LSOA.covariates}
\title{get LSOA-level covariates from the alan-turing-institute/ukhsa-turing-rss-wastewater repro}
\usage{
get.LSOA.covariates()
}
\description{
get LSOA-level covariates from the alan-turing-institute/ukhsa-turing-rss-wastewater repro
}
