% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{get.starting.day}
\alias{get.starting.day}
\title{A lookup table between time index and first day of the week}
\usage{
get.starting.day()
}
\description{
A lookup table between time index and first day of the week
}
