% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarisePrediction.R
\name{get_inset_coords}
\alias{get_inset_coords}
\title{Get the coordinate for the map inset}
\usage{
get_inset_coords(selected.region = "London", ltla.shp)
}
\description{
Get the coordinate for the map inset
}
