% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateLSOA.R
\name{get_lsoa_agg_popw}
\alias{get_lsoa_agg_popw}
\title{format data for aggregation}
\usage{
get_lsoa_agg_popw(geography, dat, aggregate.method)
}
\description{
This function carries out two operatoins: (a) constructs the lookup table between LSOA and the geography for which the aggregation is to be carried out; and (b) calculate population weights
}
