% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarisePrediction.R
\name{get_prev_sims}
\alias{get_prev_sims}
\title{Get simulated values for the debiased prevalence on the logit scale}
\usage{
get_prev_sims(nsims, lad21cd)
}
\description{
Get simulated values for the debiased prevalence on the logit scale
}
