% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{get_sites_geo}
\alias{get_sites_geo}
\title{Getting locations of the STW sites}
\usage{
get_sites_geo(site_uniq)
}
\description{
https://uwwtd.eu/United-Kingdom/download (choose csv under Title Urban Waste Water Treatment plants)
}
