% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{inCircle}
\alias{inCircle}
\title{To check if a point (x, y) is within a circle centred at (c1, c2) with radius of r}
\usage{
inCircle(c1, c2, r, x, y, plot = FALSE)
}
\description{
To check if a point (x, y) is within a circle centred at (c1, c2) with radius of r
}
