% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingData.R
\name{LSOA.to.catchment.covariates}
\alias{LSOA.to.catchment.covariates}
\title{get catchment level covariates from LSOA}
\usage{
LSOA.to.catchment.covariates(
  X_names = c("IMD_score", "all_ages"),
  op = c("mean", "sum"),
  lsoa2catchment.lookup
)
}
\description{
get catchment level covariates from LSOA
}
