% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{LSOA.to.catchment.lookup}
\alias{LSOA.to.catchment.lookup}
\title{aggregating LSOA covariates to STW catchment using the open source
lookup table https://github.com/alan-turing-institute/ukhsa-turing-rss-wastewater/tree/hoffmann-data/data/wastewater_catchment_areas_public}
\usage{
LSOA.to.catchment.lookup()
}
\description{
aggregating LSOA covariates to STW catchment using the open source
lookup table https://github.com/alan-turing-institute/ukhsa-turing-rss-wastewater/tree/hoffmann-data/data/wastewater_catchment_areas_public
}
