% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_specification.R
\name{pcprior.rho}
\alias{pcprior.rho}
\title{Setting PC prior on the AR1 coefficient that Pr(|rho|>a)=pb
By default, a=0.1 and pb=0.9 such that the prior assumes Pr(|rho|>0.1)=0.9}
\usage{
pcprior.rho(prior = "pc.cor0", a = 0.1, pb = 0.9)
}
\description{
Setting PC prior on the AR1 coefficient that Pr(|rho|>a)=pb
By default, a=0.1 and pb=0.9 such that the prior assumes Pr(|rho|>0.1)=0.9
}
