% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{read_LSOA_centroids}
\alias{read_LSOA_centroids}
\title{Read the publicly available population-weighted LSOA centroids for England}
\usage{
read_LSOA_centroids(england.only = TRUE)
}
\description{
Read the publicly available population-weighted LSOA centroids for England
}
