% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{read_public_data}
\alias{read_public_data}
\title{Read the publicly available data into R}
\usage{
read_public_data(upto = "30Mar2022")
}
\value{
A wide-format wastewater data matrix (site by day)
}
\description{
This function reads in the public wastewater data (already included in the package)
}
\details{
The data can be donwloaded via https://www.gov.uk/government/publications/monitoring-of-sars-cov-2-rna-in-england-wastewater-monthly-statistics-15-july-2020-to-30-march-2022
}
