% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingData.R
\name{replace_LOD}
\alias{replace_LOD}
\title{todo: L differs by type of measurement technique. Need to accommodate that}
\usage{
replace_LOD(
  ww,
  raw_gc_column = "gc",
  LOD.means,
  LOD.sds = NULL,
  dist = "Uniform",
  isim_LOD = NULL,
  LOD_holder = "tLOD"
)
}
\description{
If you want to replace all LODs with a fixed value (L/2 say), set
LOD.means <- L/2 and leave the rest as default
}
\details{
If you want to replace each LOD with a value randomly generated from U(0,L), set
LOD.means <- L, LOD.sds <- 0 and leave dist as default
}
