% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_specification.R
\name{set_priors}
\alias{set_priors}
\title{A function to set prior specification (apart from the AR1 coefficient; for that, see pcprior.rho)}
\usage{
set_priors(
  fixed = list(mean.intercept = 0, prec.intercept = 1e-04, mean.fixed = 0, prec.fixed =
    1e-04),
  prec.overall.time = c(10, 0.05),
  prec.overall.space = c(10, 0.05),
  prec.overall.region = c(10, 0.05),
  matern = list(range = c(10, 0.05), sigma = c(10, 0.05))
)
}
\description{
See https://github.com/gqlNU/publicWW/issues/42 for the discussion of prior specification.
The default prior is used for the residual variance, a Gamma prior with 1 and 0.00005 on the Precision
(see inla.doc("gaussian") and also https://groups.google.com/g/r-inla-discussion-group/c/MB53BG4WP68).
}
