% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingData.R
\name{ww_wide_to_long}
\alias{ww_wide_to_long}
\title{Transform the raw public wastewater data from wide to long format}
\usage{
ww_wide_to_long(dat, site_column, show.warning = TRUE)
}
\arguments{
\item{dat}{The dataset from the function read_public_data()}
}
\value{
A data matrix
}
\description{
week is the week of the year and is obtained using the ISOweek function
in the package of the same name.
}
\details{
For 01/01/2022 (a Saturday), it is allocated to week 52 in 2021.
However, day, indicating the day of the year, is 1 and month, the month of the year, is 1.
If the day (or the month) column is
used as time_index for the analysis, care needs to be taken!
For different date definitions, see https://stackoverflow.com/questions/45549449/transform-year-week-to-date-object/45587644#45587644
}
